Simplified `AccountSettings` parsing
