+++ packages/functions/src/lib/initServices.func.ts
@@ -14,10 +14,7 @@ import {prefixErrorResult} from '@shared/lib/errorUtils.shared';
 import {makeSuccessResult} from '@shared/lib/results.shared';
 
 import {parseAccount, parseAccountId} from '@shared/parsers/accounts.parser';
-import {
-  parseAccountSettings,
-  toStorageAccountSettings,
-} from '@shared/parsers/accountSettings.parser';
+import {parseAccountSettings} from '@shared/parsers/accountSettings.parser';
 import {parseEventId, parseEventLogItem} from '@shared/parsers/eventLog.parser';
 import {
   parseAccountExperimentsState,
@@ -35,6 +32,7 @@ import type {Result} from '@shared/types/results.types';
 import type {RssFeedProvider} from '@shared/types/rss.types';
 
 import {toStorageAccount} from '@shared/storage/accounts.storage';
+import {toStorageAccountSettings} from '@shared/storage/accountSettings.storage';
 import {toStorageEventLogItem} from '@shared/storage/eventLog.storage';
 
 import {ServerAccountsService} from '@sharedServer/services/accounts.server';
